requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-filter==2.4.0
django-stubs==1.7.0
djangorestframework==3.12.2
djangorestframework-stubs==1.3.0
drf-access-policy==0.8.1
drf-extensions==0.6.0
drf-flex-fields==0.8.8
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.4
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted @ file:///C:/Users/TheNightmareX/Desktop/ChaaAt_backend/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.4.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
